refactor: move TotalBalance component to atoms folder
