+++ src/components/atoms/index.ts
@@ -14,3 +14,4 @@ export { Switch } from './switch';
 export { BottomSheetTokenItem } from './bottom-sheet-token-item';
 export { QrScannerDebugger } from './qr-scanner-debugger';
 export * from './input';
+export * from './total-balance';
